+++ src/nbrefactor/__init__.py
@@ -1,5 +1,5 @@
 # Current
-__version__ = '0.1.3'
+__version__ = '0.1.4'
 
 # CLI
 from .cli import main